server/requirement_vosk.txt
vosk
pyaudio
pydub
pyyaml
numpy
sounddevice
Flask